ASoC: q6asm: Fix the memory leak during the SSR
If a session is released during the SSR, then the memory unmap
command to DSP fails. This results in not releasing up the memory
block shared among apps and the DSP.
Fix by checking if the session has been invalidated during the SSR
and free up the memory during the close command.

CRs-Fixed: 687190
Change-Id: I66251e624e5cc1abd816c2e4647997482b96c771
Signed-off-by: Aviral Gupta <aviralg@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>