ASoC: wcd9xxx: modify sequence post SSR
Perform mux configuration before MBHC init.
Otherwise, MLCK will not be enabled which results
in wrong initial calibration of dce and sta.
Updating the mux configuration for class h dsm mux
during codec bring up after ssr is causing the
widget not to be powered up as the dapm sees no
change in the register associated with it. The change
ensures the mux register is not updated post SSR.

Change-Id: I4d181de29816d51b0a70b82bd057256844d34387
Signed-off-by: Aravind Kumar <akumark@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>