usb: dwc3: gadget: Ignore L1 RESUME events
Events may get generated for Link Power Management L1
Suspend and Resume. The suspend case currently ignores
L1 and only proceeds to call the gadget driver suspend
routines if it is for L2/U3 entry. However, the resume
callback doesn't make this distinction, which will
result in the gadget layer resume getting called for
which there had not been a previous suspend.

Change-Id: I9c0e866d9c834a8ea7b3f52e65fa702efbee2b38
CRs-Fixed: 739812
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>