ASoC: msm: qdsp6v2: Unblock EOS on reset event callback
Issue: When SSR command is issued in the end of stream of
compress offload playback, ANR is seen in music app. This
is because EOS is waiting indefinitely and will not return
since ADSP is dead in this case.

Fix is to unblock EOS on reset event callback from DSP.

Change-Id: Ifc379360b0ba7c589c3e2dddb7b2a3c712af7fb0
CRs-Fixed: 731606
Signed-off-by: Divya Narayanan Poojary <dnaray@codeaurora.org>
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>
Signed-off-by: Banajit Goswami <bgoswami@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>